Move `MigrateScheme` to new Onceoff thing
